Ping later.
Ping can timeout (and take a very long time doing so), meaning that we have to
wait a good while before we get the information that isn't delayed in this
way.

Move the command until later to mitigate this. It sits immediately before
commands that can fail completely like rce images and rce ps -a, so this is as
far down the list as the command should sit.