Pass script to leech via stdin.
This avoids establishing 2 separate connections in order to grab diagnostics
which significantly improves time taken and avoids creating a mess on the
device's filesystem.

Arguably copying then executing is more reliable as connection interruptions
might impact script execution, however since we're simply taking output from
stdout we are clearly not utilising this so it makes no sense to retain this
method.