Remove image corruption check, as it is currently too noisy
This check in its current form has a number of disadvantages:

- It causes a lot of IO.  Under the hood, balena-engine is:

  - reading the individual layers;
  - writing them out to /var/lib/docker/tmp;
  - then reading them all back in, so it can write to stdout.

- One consequence is that the `docker save` command can take longer
than the current 10s timeout (particularly if there are a lot of
images, or if the images are large), despite the device and its
storage being perfectly healthy.

- While it is good to know if an image has corruption, it's probably
more important to know if the underlying storage is having problems;
this check is not a great way of determining if that is happening.

Thus, for now we're better off without this check.  We can investigate
other approaches to verifying images, or examining storage for problems.

Connects-to: #203
Change-type: minor
Signed-off-by: Hugh Brown <hugh@balena.io>