diagnose.sh: fix timeout 'command not found' error
The dashboard 'Standard error' display for 'Device Diagnostics'
reports the following error:
'bash: line 27: timeout -v 1: command not found'

This issue is due to the double quotes used in the corresponding
'if' statement. Adding an additional redirect of stderr to this
commands hides any error messages for the cases where the '-v'
option is not supported by the version of 'timeout' being used.

This fix has been tested as a simple, isolated code snippet on the
balenaOS (2.51.1+rev1) command line running on a RPi3.

Change-type: patch
Connects-to: #234
Signed-off-by: Mark Corbin <mark@balena.io>